+++ main.tf
@@ -1,42 +1,6 @@
-provider "aws" {
-  alias = "master"
-}
-
 # Enable SecurityHub
 module "account" {
   source = "./modules/account"
-}
-
-# Send invite from master
-module "member" {
-  source = "./modules/member"
-  count  = local.cross_account ? 1 : 0
-  providers = {
-    aws = aws.master
-  }
-
-  account_id = data.aws_caller_identity.this.account_id
-  email      = var.member_email
-
-  depends_on = [
-    module.account
-  ]
-}
-
-# Accept invite
-module "accept" {
-  source = "./modules/accepter"
-  count  = local.cross_account ? 1 : 0
-
-  master_account_id = data.aws_caller_identity.master.account_id
-
-  profile  = var.accepter_profile
-  role_arn = var.accepter_role_arn
-  region   = var.accepter_region
-
-  depends_on = [
-    module.member
-  ]
 }
 
 # Manage subscriptions
@@ -59,14 +23,4 @@ module "action_targets" {
   name        = each.key
   description = each.value.description
   identifier  = each.value.identifier
-}
-
-locals {
-  cross_account = data.aws_caller_identity.this.account_id != data.aws_caller_identity.master.account_id
-}
-
-data "aws_caller_identity" "this" {}
-
-data "aws_caller_identity" "master" {
-  provider = aws.master
 }