+++ tests/securityhub_cross_account/main.tf
@@ -9,17 +9,15 @@ provider "aws" {
   profile = "resource-owner"
 }
 
-module "securityhub_member" {
-  source = "../../"
+module "securityhub" {
+  source = "../../modules/cross-account-member"
 
   providers = {
-    aws        = aws
-    aws.master = aws.resource-owner
+    aws               = aws
+    aws.administrator = aws.resource-owner
   }
 
-  member_email     = var.member_email
-  accepter_region  = data.aws_region.current.name
-  accepter_profile = "resource-member"
+  member_email = var.member_email
 
   standard_subscription_arns = [
     "arn:aws:securityhub:::ruleset/cis-aws-foundations-benchmark/v/1.2.0",
@@ -31,8 +29,6 @@ module "securityhub_member" {
   ]
 }
 
-data "aws_region" "current" {}
-
 output "securityhub" {
-  value = module.securityhub_member
+  value = module.securityhub
 }